+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## 1.0.5
+  * When loading a file via `pkgutil` it must be decoded from bytes [#57](https://github.com/singer-io/tap-google-analytics/pull/56)
+
+## 1.0.4
+  * Fix inclusion of `ga_cubes.json` and remove dynamic pulling of the file for reliability [#56](https://github.com/singer-io/tap-google-analytics/pull/56)
+
+## [v1.0.2](https://github.com/singer-io/tap-google-analytics/tree/v1.0.2) (2021-01-05)
+
+[Full Changelog](https://github.com/singer-io/tap-google-analytics/compare/v1.0.1...v1.0.2)
+
+**Merged pull requests:**
+
+- Properly retry on 503 and add unit tests for all status codes that should retry under all conditions [\#43](https://github.com/singer-io/tap-google-analytics/pull/43) ([asaf-erlich](https://github.com/asaf-erlich))
+- move time off of utc midnight [\#41](https://github.com/singer-io/tap-google-analytics/pull/41) ([kspeer825](https://github.com/kspeer825))
+
+## 1.0.3
+  * Wraps access token request in a requests session [#45](https://github.com/singer-io/tap-google-analytics/pull/45)
+
+## 1.0.1
+  * Grabs error reason from 'reason', 'error_description', or falls back to full json response to ensure a good error message [#39](https://github.com/singer-io/tap-google-analytics/pull/39)
+  * add context user [#38](https://github.com/singer-io/tap-google-analytics/pull/38)
+
 ## 1.0.0
   * Releasing to GA